+++ BackEnd/loginservice/src/main/java/com/banking/project/loginservice/services/MyUserDetails.java
@@ -19,7 +19,7 @@ public class MyUserDetails implements UserDetails {
 	private String username;
 	private String firstName;
 	private String lastName;
-	private int dateOfBirth;
+	private long dateOfBirth;
 	private String gender;
 	@JsonIgnore
 	private String password;
@@ -27,7 +27,7 @@ public class MyUserDetails implements UserDetails {
 	private SimpleGrantedAuthority authority;
 
 	
-	public MyUserDetails(int id, String username, String password,String firstName, String lastName, int dateOfBirth, String gender,
+	public MyUserDetails(int id, String username, String password,String firstName, String lastName, long dateOfBirth, String gender,
 			SimpleGrantedAuthority authority) {
 		this.id = id;
 		this.username = username;
@@ -81,7 +81,7 @@ public class MyUserDetails implements UserDetails {
 		return lastName;
 	}
 
-	public int getDateOfBirth() {
+	public long getDateOfBirth() {
 		return dateOfBirth;
 	}
